#277 - migrating dialogfields
