+++ bb-aem-core/src/main/java/com/cognifide/qa/bb/aem/core/component/dialog/dialogfields/text/ControlToolbar.java
@@ -1,6 +1,9 @@
-/*
- * Copyright 2016 Cognifide Ltd..
- *
+/*-
+ * #%L
+ * Bobcat
+ * %%
+ * Copyright (C) 2018 Cognifide Ltd.
+ * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -12,78 +15,31 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ * #L%
  */
 package com.cognifide.qa.bb.aem.core.component.dialog.dialogfields.text;
 
-import com.cognifide.qa.bb.qualifier.CurrentScope;
-import com.cognifide.qa.bb.qualifier.PageObject;
-import com.google.inject.Inject;
-import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.interactions.Actions;
-import org.openqa.selenium.support.FindBy;
 
 /**
- * Represents RichTexts controls toolbar.
+ * Interface that represents RichTexts controls toolbar.
  */
-@PageObject
-public class ControlToolbar {
-
-  static final String TOOLBAR_ITEM_SELECTOR = ".coral-RichText-toolbar-item";
-
-  @Inject
-  private Actions actions;
-
-  @Inject
-  @CurrentScope
-  private WebElement scope;
-
-  @FindBy(css = TOOLBAR_ITEM_SELECTOR + "[data-action='format#bold']")
-  private WebElement toggleBoldButton;
-
-  @FindBy(css = TOOLBAR_ITEM_SELECTOR + "[data-action='format#italic']")
-  private WebElement toggleItalicButton;
-
-  @FindBy(css = TOOLBAR_ITEM_SELECTOR + "[data-action='format#underline']")
-  private WebElement toggleUnderlineButton;
-
-  @FindBy(css = TOOLBAR_ITEM_SELECTOR + "[data-action='#justify']")
-  private WebElement toggleJustifyButton;
-
-  @FindBy(css = TOOLBAR_ITEM_SELECTOR + "[data-action='#lists']")
-  private WebElement toggleListButton;
+public interface ControlToolbar {
 
   /**
    * Performs 'select all text' action.
    */
-  public void selectText() {
-    actions.keyDown(Keys.CONTROL)
-            .sendKeys("a")
-            .perform();
-  }
+  void selectText();
 
-  public WebElement getScope() {
-    return scope;
-  }
+  WebElement getScope();
 
-  public WebElement getToggleBoldButton() {
-    return toggleBoldButton;
-  }
+  WebElement getToggleBoldButton();
 
-  public WebElement getToggleItalicButton() {
-    return toggleItalicButton;
-  }
+  WebElement getToggleItalicButton();
 
-  public WebElement getToggleUnderlineButton() {
-    return toggleUnderlineButton;
-  }
+  WebElement getToggleUnderlineButton();
 
-  public WebElement getToggleJustifyButton() {
-    return toggleJustifyButton;
-  }
+  WebElement getToggleJustifyButton();
 
-  public WebElement getToggleListButton() {
-    return toggleListButton;
-  }
-
+  WebElement getToggleListButton();
 }
